[expo-notifications] Extract Base64 serialization to a separate file
