+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/service/delegates/SharedPreferencesNotificationsStore.kt
@@ -15,7 +15,6 @@ import java.io.ObjectOutputStream
  * A fairly straightforward [SharedPreferences] wrapper to be used by [NotificationSchedulingHelper].
  * Saves and reads notifications (identifiers, requests and triggers) to and from the persistent storage.
  *
- *
  * A notification request of identifier = 123abc, it will be persisted under key:
  * [SharedPreferencesNotificationsStore.NOTIFICATION_REQUEST_KEY_PREFIX]123abc
  */
@@ -38,16 +37,13 @@ class SharedPreferencesNotificationsStore(context: Context) {
    */
   @Throws(IOException::class, ClassNotFoundException::class)
   fun getNotificationRequest(identifier: String) =
-    deserializeNotificationRequest(
-      sharedPreferences.getString(
-        preferencesNotificationRequestKey(identifier),
-        null
-      )
-    )
+    sharedPreferences.getString(
+      preferencesNotificationRequestKey(identifier),
+      null
+    )?.asBase64EncodedObject<NotificationRequest>()
 
   /**
    * Fetches all scheduled notifications, ignoring invalid ones.
-   *
    *
    * Goes through all the [SharedPreferences] entries, interpreting only the ones conforming
    * to the expected format.
@@ -56,18 +52,20 @@ class SharedPreferencesNotificationsStore(context: Context) {
    */
   val allNotificationRequests: Collection<NotificationRequest>
     get() =
-      sharedPreferences.all.mapNotNull { (key, value) ->
-        try {
-          if (key.startsWith(NOTIFICATION_REQUEST_KEY_PREFIX)) {
-            return@mapNotNull deserializeNotificationRequest(value as String?)
+      sharedPreferences
+        .all
+        .filter { it.key.startsWith(NOTIFICATION_REQUEST_KEY_PREFIX) }
+        .mapNotNull { (_, value) ->
+          return@mapNotNull try {
+            (value as String?)?.asBase64EncodedObject<NotificationRequest>()
+          } catch (e: ClassNotFoundException) {
+            // do nothing
+            null
+          } catch (e: IOException) {
+            // do nothing
+            null
           }
-        } catch (e: ClassNotFoundException) {
-          // do nothing
-        } catch (e: IOException) {
-          // do nothing
         }
-        null
-      }
 
   /**
    * Saves given notification in the persistent storage.
@@ -80,7 +78,7 @@ class SharedPreferencesNotificationsStore(context: Context) {
     sharedPreferences.edit()
       .putString(
         preferencesNotificationRequestKey(notificationRequest.identifier),
-        serializeNotificationRequest(notificationRequest)
+        notificationRequest.encodedInBase64()
       )
       .apply()
 
@@ -116,45 +114,6 @@ class SharedPreferencesNotificationsStore(context: Context) {
       }.let {
         this.apply()
         it
-      }
-    }
-
-  /**
-   * Serializes the trigger to a Base64-encoded string.
-   *
-   * @param notificationRequest Notification request to serialize
-   * @return Base64-encoded, serialized trigger
-   * @throws IOException Thrown if there is an error while writing trigger to string.
-   */
-  @Throws(IOException::class)
-  private fun serializeNotificationRequest(notificationRequest: NotificationRequest) =
-    ByteArrayOutputStream().use { byteArrayOutputStream ->
-      ObjectOutputStream(byteArrayOutputStream).use { objectOutputStream ->
-        objectOutputStream.writeObject(notificationRequest)
-        Base64.encodeToString(byteArrayOutputStream.toByteArray(), Base64.NO_WRAP)
-      }
-    }
-
-  /**
-   * Deserializes trigger from the string representation.
-   *
-   * @param trigger Base64-encoded, serialized trigger representation
-   * @return Deserialized trigger
-   * @throws IOException            Thrown if there is an error while reading trigger from String
-   * @throws ClassNotFoundException Thrown if the deserialization failes due to class not being found.
-   * @throws InvalidClassException  Thrown if the trigger is of invalid class.
-   */
-  @Throws(IOException::class, ClassNotFoundException::class, InvalidClassException::class)
-  private fun deserializeNotificationRequest(trigger: String?): NotificationRequest =
-    Base64.decode(trigger, Base64.NO_WRAP).let {
-      ByteArrayInputStream(it).use { byteArrayInputStream ->
-        ObjectInputStream(byteArrayInputStream).use { ois ->
-          val o = ois.readObject()
-          if (o is NotificationRequest) {
-            return o
-          }
-          throw InvalidClassException("Expected serialized notification request to be an instance of NotificationRequest. Found: $o")
-        }
       }
     }
